Feature/ms bank account

+++ bank-account/src/main/java/com/dac/bank_account/command/dto/request/MovementRequestDTO.java
@@ -8,6 +8,6 @@ import java.math.BigDecimal;
 public record MovementRequestDTO(
         @NotNull
         @Positive
-        BigDecimal valor
+        Double valor
 ) {
 }
