+++ bank-account/src/main/java/com/dac/bank_account/command/dto/request/TransferRequestDTO.java
@@ -11,6 +11,6 @@ public record TransferRequestDTO(
         String destino,
         @NotNull
         @Positive
-        BigDecimal valor
+        Double valor
 ) {
 }
