+++ bank-account/src/main/java/com/dac/bank_account/command/dto/response/AccountResponseDTO.java
@@ -1,13 +1,10 @@
 package com.dac.bank_account.command.dto.response;
 
-
-import java.math.BigDecimal;
-
 public record AccountResponseDTO(
         String cliente,
         String numero,
-        BigDecimal saldo,
-        BigDecimal limite,
+        Double saldo,
+        Double limite,
         String gerente,
         String criacao
 ) {
