+++ bank-account/src/main/java/com/dac/bank_account/query/dto/BalanceResponseDTO.java
@@ -5,5 +5,5 @@ import java.math.BigDecimal;
 public record BalanceResponseDTO(
         String cliente,
         String conta,
-        BigDecimal saldo
+        Double saldo
 ) {}
