+++ bank-account/src/main/java/com/dac/bank_account/query/dto/StatementResponseDTO.java
@@ -5,6 +5,6 @@ import java.util.List;
 
 public record StatementResponseDTO(
         String conta,
-        BigDecimal saldo,
+        Double saldo,
         List<TransactionResponseDTO> movimentacoes
 ) {}
