+++ bank-account/src/main/java/com/dac/bank_account/query/dto/TransactionResponseDTO.java
@@ -7,6 +7,6 @@ public record TransactionResponseDTO(
         String tipo,
         String origem,
         String destino,
-        BigDecimal valor
+        Double valor
 ) {
 }
